Update more unit tests. IT tests still failing.

+++ data-query/src/main/java/gov/va/api/health/dataquery/service/controller/medicationrequest/R4MedicationRequestController.java
@@ -260,7 +260,7 @@ public class R4MedicationRequestController {
         medicationStatementRepository.count(medicationStatementPatientSpec);
 
     int lastPageWithMedicationStatement =
-        (int) Math.floor(numMedicationStatements / (count == 0 ? 1 : count));
+        (int) Math.floor((double) numMedicationStatements / (count == 0 ? 1 : count));
 
     Page<MedicationStatementEntity> medicationStatementEntities =
         medicationStatementRepository.findByIcn(
